Split tests and remove duplicated tests
