+++ packages/npm-packages/ruby-wasm-wasi/test/eval.test.ts
@@ -1,0 +1,9 @@
+import { initRubyVM } from "./init";
+
+describe("Ruby code evaluation", () => {
+  test("empty expression", async () => {
+    const vm = await initRubyVM();
+    const result = vm.eval("");
+    expect(result.toString()).toBe("");
+  });
+});
