+++ packages/npm-packages/ruby-wasm-wasi/test/test_unit.rb
@@ -1,5 +1,7 @@
 require_relative "./unit/test_js"
 require_relative "./unit/test_object"
+require_relative "./unit/test_object_wrap"
 require_relative "./unit/test_error"
 require_relative "./unit/test_async"
 require_relative "./unit/test_float"
+require_relative "./unit/test_proc"
