+++ packages/npm-packages/ruby-wasm-wasi/test/vm.test.ts
@@ -1,4 +1,3 @@
-import { RbValue } from "../src/index";
 import { initRubyVM } from "./init";
 
 describe("RubyVM", () => {
@@ -69,42 +68,6 @@ describe("RubyVM", () => {
     }).toThrowError("Can't escape from eval with redo");
   });
 
-  test("protect exported Ruby objects", async () => {
-    function dropRbValue(value: RbValue) {
-      (value as any).inner.drop();
-    }
-    const vm = await initRubyVM();
-    const initialGCCount = Number(vm.eval("GC.count").toString());
-    const robj = vm.eval("$x = Object.new");
-    const robjId = robj.call("object_id").toString();
-    expect(robjId).not.toEqual("");
-
-    vm.eval("GC.start");
-    expect(robj.call("object_id").toString()).toBe(robjId);
-    expect(Number(vm.eval("GC.count").toString())).toEqual(initialGCCount + 1);
-
-    const robj2 = vm.eval("$x");
-    vm.eval("GC.start");
-    expect(robj2.call("object_id").toString()).toBe(robjId);
-
-    const robj3 = robj2.call("itself");
-    vm.eval("GC.start");
-    expect(robj3.call("object_id").toString()).toBe(robjId);
-
-    dropRbValue(robj);
-    expect(robj2.call("object_id").toString()).toBe(robjId);
-    expect(robj3.call("object_id").toString()).toBe(robjId);
-
-    vm.eval("GC.start");
-    expect(robj2.call("object_id").toString()).toBe(robjId);
-    expect(robj3.call("object_id").toString()).toBe(robjId);
-
-    dropRbValue(robj2);
-    dropRbValue(robj3);
-
-    vm.eval("GC.start");
-  });
-
   test("method call with args", async () => {
     const vm = await initRubyVM();
     const X = vm.eval(`
@@ -148,48 +111,6 @@ describe("RubyVM", () => {
 eval:6:in \`bar'
 eval:3:in \`foo'
 eval:11:in \`<main>'`);
-  });
-
-  test("protect objects having same hash values from GC", async () => {
-    const vm = await initRubyVM();
-    vm.eval(`
-    class X
-      def hash
-        42
-      end
-      def eql?(other)
-        true
-      end
-    end
-    `);
-
-    const o1 = vm.eval(`X.new`);
-    const o2 = vm.eval(`X.new`);
-    const o3 = vm.eval(`X.new`);
-    vm.eval(`GC.start`);
-    expect(o1.call("hash").toString()).toBe(o2.call("hash").toString());
-    expect(o2.call("hash").toString()).toBe(o3.call("hash").toString());
-  });
-
-  test("Wrap arbitrary JS object to RbValue", async () => {
-    const vm = await initRubyVM();
-    const o1 = {
-      v() {
-        return 42;
-      },
-    };
-    const X = vm.eval(`
-    module X
-      def self.identity(x) = x
-    end
-    X
-    `);
-    const o1Clone = X.call("identity", vm.wrap(o1));
-    expect(o1Clone.call("call", vm.eval(`"v"`)).toJS().toString()).toBe("42");
-
-    // Check that JS object can be stored in Ruby Hash
-    const hash = vm.eval(`Hash.new`);
-    hash.call("store", vm.eval(`"key1"`), vm.wrap(new Object()));
   });
 
   test("eval encoding", async () => {
